+++ src/SumatraPDF.cpp
@@ -6773,7 +6773,7 @@ static LRESULT FrameOnCommand(MainWindow* win, HWND hwnd, UINT msg, WPARAM wp, L
             ShowPdfBakeDialog(win);
             break;
 
-        case CmdPdfExtractText:
+        case CmdDocumentExtractText:
             ShowPdfExtractTextDialog(win);
             break;